Change UserPreferences to RestaurantFilteringController
